Support scopes with parameters
That's a significant change and changes the API surface. kotlin-inject supports scopes with and
without parameters. This change implements a similar API as Anvil for Dagger 2 used to provide,
where scope references can be added to annotations, e.g.
```kotlin
@ContributesTo(AppScope::class)
interface ContributedComponentInterface

@Component
@MergeComponent(AppScope::class)
interface MergedComponent
```

The existing mechanism by annotating classes with the scope annotation is still working:
```kotlin
@ContributesTo
@SingleInAppScope
interface ContributedComponentInterface

@Component
@MergeComponent
@SingleInAppScope
interface MergedComponent
```

Most of the logic changed in the way we resolve the scope where to merge code, which happens during compilation.

There is an additional change for generated code, where we no longer add the scope to generated interfaces and
instead rely on the `@Origin` annotation to determine the scope. That simplifies code generation and is backwards
compatible.

Resolves #1.