Provide option to hide the merged interface
Until now it was required to add the merged interface as super type, e.g.
```
@Component
@MergeComponent(AppScope::class)
interface AppComponent : AppComponentMerged
```

With the new mechanism the `@Component` annotation and the super type can be omitted. In this case we will generate the final kotlin-inject component under the hood. This removes boilerplate and brings us closer to the original Anvil design. Further, this will help with #20 in KMP scenarios where generated code cannot be access from common code.

Fixes #8